crawler/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "1group-cuisine-crawler"
version = "0.1.0"
description = "Professional culinary crawler for 1-Group Cuisine platform"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "1-Group Cuisine Team", email = "tech@1group-cuisine.com"},
]
license = {text = "Proprietary"}
dependencies = [
    "scrapy>=2.11.0",
    "playwright>=1.40.0",
    "httpx>=0.25.0",
    "anthropic>=0.7.0",
    "extruct>=0.15.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "sentence-transformers>=2.2.0",
    "Pillow>=10.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "asyncpg>=0.29.0",
    "structlog>=23.2.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
]

[project.scripts]
1group-crawler = "crawler.cli:main"

[tool.setuptools]
packages = ["crawler"]

[tool.setuptools.package-data]
crawler = ["config/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
